Changing to float32 without to_gpu conversion

+++ parameters.py
@@ -102,38 +102,38 @@ def make_weights_and_masks():
 
 	# Make W_in and mask
 	par['W_in_init'] = np.random.gamma(par['input_gamma'], \
-		scale=1., size=[par['n_input'], par['n_hidden']])
+		scale=1., size=[par['n_input'], par['n_hidden']]).astype(np.float32)
 	par['W_in_mask'] = np.zeros_like(par['W_in_init'])
 	par['W_in_mask'][:,0:par['n_EI']:4] = 1.
 	par['W_in_init'] *= par['W_in_mask']
 
 	# Make W_out and mask
-	par['W_out_init'] = np.random.uniform(-1., 1., size=[par['n_hidden'], par['n_output']])
+	par['W_out_init'] = np.random.uniform(-1., 1., size=[par['n_hidden'], par['n_output']]).astype(np.float32)
 	par['W_out_mask'] = np.ones_like(par['W_out_init'])
 
 	# Make b_out and mask
-	par['b_out_init'] = np.zeros([1, par['n_output']])
+	par['b_out_init'] = np.zeros([1, par['n_output']], dtype=np.float32)
 	par['b_out_mask'] = np.ones_like(par['b_out_init'])
 
 	# Make W_rnn and mask
 	if par['EI_prop'] == 1.:
-		par['W_rnn_init'] = np.random.uniform(-par['rnn_gamma'], par['rnn_gamma'], size=[par['n_hidden'],par['n_hidden']])
+		par['W_rnn_init'] = np.random.uniform(-par['rnn_gamma'], par['rnn_gamma'], size=[par['n_hidden'],par['n_hidden']]).astype(np.float32)
 	else:
-		par['W_rnn_init'] = np.random.gamma(par['rnn_gamma'], scale=1.0, size=[par['n_hidden'], par['n_hidden']])
+		par['W_rnn_init'] = np.random.gamma(par['rnn_gamma'], scale=1.0, size=[par['n_hidden'], par['n_hidden']]).astype(np.float32)
 		if par['balance_EI']:
 			par['W_rnn_init'][par['n_EI']:,:] *= 1.8
 			par['W_rnn_init'][:,par['n_EI']:] *= 1.8
 
-	par['W_rnn_mask']   = 1 - np.eye(par['n_hidden'])
+	par['W_rnn_mask']   = 1 - np.eye(par['n_hidden']).astype(np.float32)
 	par['W_rnn_init']  *= par['W_rnn_mask']
 
-	par['W_rnn_init'] = np.minimum(2., par['W_rnn_init'])
+	par['W_rnn_init'] = np.minimum(2., par['W_rnn_init']).astype(np.float32)
 
 	# Remake W_in and mask if weight won't be trained
 	if not par['train_input_weights']:
 
-		par['W_in_const'] = np.zeros_like(par['W_in_init'])
-		U = np.linspace(0, 360, par['n_input'])
+		par['W_in_const'] = np.zeros_like(par['W_in_init'], dtype=np.float32)
+		U = np.linspace(0, 360, par['n_input']).astype(np.float32)
 
 		beta = 0.1
 		kappa = 1.
@@ -178,17 +178,17 @@ def update_dependencies():
 	par['n_EI'] = int(par['n_hidden']*par['EI_prop'])
 
 	# Generate EI vector and matrix
-	par['EI_vector'] = np.ones(par['n_hidden'])
+	par['EI_vector'] = np.ones(par['n_hidden'], dtype=np.float32)
 	par['EI_vector'][par['n_EI']:] *= -1
-	par['EI_matrix'] = np.diag(par['EI_vector'])
+	par['EI_matrix'] = np.diag(par['EI_vector']).astype(np.float32)
 
-	par['exh_vector'] = np.ones(par['n_hidden'])
+	par['exh_vector'] = np.ones(par['n_hidden'], dtype=np.float32)
 	par['exh_vector'][par['n_EI']:] *= 0
-	par['EI_mask_exh'] = np.diag(par['exh_vector'])
+	par['EI_mask_exh'] = np.diag(par['exh_vector']).astype(np.float32)
 
-	par['inh_vector'] = np.ones(par['n_hidden'])
+	par['inh_vector'] = np.ones(par['n_hidden'], dtype=np.float32)
 	par['inh_vector'][:par['n_EI']] *= 0
-	par['EI_mask_inh'] = np.diag(par['inh_vector'])
+	par['EI_mask_inh'] = np.diag(par['inh_vector']).astype(np.float32)
 
 	# Initialize weights and generate the associated masks
 	make_weights_and_masks()
@@ -212,12 +212,12 @@ def update_dependencies():
 	### STP
 	if par['use_stp']:
 
-		par['alpha_stf'] = np.ones([1,par['n_hidden'],1])
-		par['alpha_std'] = np.ones([1,par['n_hidden'],1])
-		par['U']         = np.ones([1,par['n_hidden'],1])
+		par['alpha_stf'] = np.ones([1,par['n_hidden'],1], dtype=np.float32)
+		par['alpha_std'] = np.ones([1,par['n_hidden'],1], dtype=np.float32)
+		par['U']         = np.ones([1,par['n_hidden'],1], dtype=np.float32)
 
-		par['syn_x_init'] = np.zeros([par['batch_size'],par['n_hidden'],1])
-		par['syn_u_init'] = np.zeros([par['batch_size'],par['n_hidden'],1])
+		par['syn_x_init'] = np.zeros([par['batch_size'],par['n_hidden'],1], dtype=np.float32)
+		par['syn_u_init'] = np.zeros([par['batch_size'],par['n_hidden'],1], dtype=np.float32)
 
 		for i in range(0,par['n_hidden'],2):
 			par['alpha_stf'][:,i,:] = par['dt']/par['tau_slow']
@@ -252,7 +252,7 @@ def update_dependencies():
 
 		for (k0, v_exc), (k1, v_inh) in zip(par[par['exc_model']].items(), par[par['inh_model']].items()):
 			assert(k0 == k1)
-			par_matrix = np.ones([1,1,par['n_hidden']])
+			par_matrix = np.ones([1,1,par['n_hidden']], dtype=np.float32)
 			par_matrix[:,:,:par['n_EI']] *= v_exc
 			par_matrix[:,:,par['n_EI']:] *= v_inh
 			par['adex'][k0] = par_matrix
@@ -262,8 +262,8 @@ def update_dependencies():
 		par['adex']['dt']  = par['dt']/1000
 		par['w_init']      = par['adex']['b']
 
-		par['adex']['beta']  = np.exp(-par['dt']/par['tau_hid'])
-		par['adex']['kappa'] = np.exp(-par['dt']/par['tau_out'])
+		par['adex']['beta']  = np.exp(-par['dt']/par['tau_hid']).astype(np.float32)
+		par['adex']['kappa'] = np.exp(-par['dt']/par['tau_out']).astype(np.float32)
 		par['adex']['mu']    = par['current_multiplier']
 
 
@@ -276,9 +276,9 @@ def update_dependencies():
 			'v_th'      : 0.61,
 			'beta'      : 1.8
 		}
-		par['lif']['alpha'] = np.exp(-par['dt_sec']/par['lif']['tau_m'])
-		par['lif']['rho']   = np.exp(-par['dt_sec']/par['lif']['tau_a'])
-		par['lif']['kappa'] = np.exp(-par['dt_sec']/par['lif']['tau_o'])
+		par['lif']['alpha'] = np.exp(-par['dt_sec']/par['lif']['tau_m']).astype(np.float32)
+		par['lif']['rho']   = np.exp(-par['dt_sec']/par['lif']['tau_a']).astype(np.float32)
+		par['lif']['kappa'] = np.exp(-par['dt_sec']/par['lif']['tau_o']).astype(np.float32)
 
 
 update_dependencies()
